+++ src/commonMain/kotlin/AutoLoggers.kt
@@ -4,3 +4,5 @@ val Any.logTag
     get() = this::class.simpleName ?: error("Unable to retrieve log tag")
 val Any.logger: KSLog
     get() = TagLogger(logTag)
+
+fun taggedLogger(tagBase: Any): KSLog = tagBase.logger